requirements.txt
aiohttp>=3.8.0
pandas>=1.5.0
solana>=0.30.2
python-dotenv>=1.0.0
requests>=2.31.0
loguru>=0.7.0
solders>=0.18.0 